% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{clean_filename}
\alias{clean_filename}
\title{clean a file name}
\usage{
clean_filename(in_file, replace_special = "-")
}
\arguments{
\item{in_file}{the file to copy}

\item{replace_special}{what to use to replace anything besides dot}
}
\value{
character
}
\description{
given a file to copy, generates a new file name stripped of special
characters
}
\details{
This function creates a new file name that is stripped of special
characters. The following characters are replaced by a period: " ", ":", "~",
"'". A leading "~" will be replaced by space.
}
\examples{
clean_filename("badly:formatted~file name(sample1)")

}
