% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{create_master_meta_data}
\alias{create_master_meta_data}
\title{create meta-data}
\usage{
create_master_meta_data(file_dirs = ".", json_meta = "all_meta_data.json",
  recursive = TRUE)
}
\arguments{
\item{file_dirs}{the directories to search}

\item{json_meta}{where to save the new file}

\item{recursive}{should they be searched recursively?}
}
\description{
Given a set of directories to search for JSON meta-data files, remake
a complete meta-data file and save it to the location specified.
}
