% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{import_json}
\alias{import_json}
\title{import json}
\usage{
import_json(json_file)
}
\arguments{
\item{json_file}{the json file to read}
}
\description{
import json from a file correctly given some things where things get written
differently
}
