% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{update_hashes}
\alias{update_hashes}
\title{update hashes}
\usage{
update_hashes(json_file)
}
\arguments{
\item{json_file}{the json file that needs to be modified}
}
\description{
If you used a previous version of this package that used MD5 hashes and now
need to fix it by changing to SHA-1 hashes, this is your function.
}
